Оформил save extension битмапа как обычный метод статического класса
